First concept for new assert front end.
I couldn't make the server work so to test the designs, just serve
static files from ./templates/. I'll see if I can set it up properly
later.

All the changes are within new files to avoid breaking anything (the file
name extention for the new files is prefixed with .edaz)

The New CSS is in the templates folder because, as I said above, I
couldn't make the server work properly. If you move it, you'll need to
update the references inside the HTML files.